Move separated features to the correct package
